Persist list state between the layouts
