+++ src/tui/layouts/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { RESULT_LIST_LAYOUT, SEARCH_LAYOUT, DETAIL_LAYOUT } from "../../constants";
-import { LayoutWrapper, Layout } from "./Layout";
+import { Layout } from "./Layout";
 import Search from "./search";
 import ResultList from "./result-list";
 import { ResultListContextProvider } from "../contexts/ResultListContext";
@@ -8,7 +8,7 @@ import Detail from "./detail";
 
 const Layouts: React.FC = () => {
   return (
-    <LayoutWrapper initialLayout={SEARCH_LAYOUT}>
+    <>
       <Layout layoutName={SEARCH_LAYOUT}>
         <Search />
       </Layout>
@@ -22,7 +22,7 @@ const Layouts: React.FC = () => {
       <Layout layoutName={DETAIL_LAYOUT}>
         <Detail />
       </Layout>
-    </LayoutWrapper>
+    </>
   );
 };
